OE-6684 Make UV pulse duration in CXL Dataset a decimal as can have fractional seconds.

+++ protected/controllers/CxlDatasetController.php
@@ -772,7 +772,7 @@ EOL;
                     RiboflavinPreparation VARCHAR(128),
                     RiboflavinDuration INTEGER,
                     UVIrradiance INTEGER,
-                    UVDuration INTEGER,
+                    UVDuration DECIMAL(5,1),
                     UVContinuousOrPulsed VARCHAR(128),
                     UVTotalEnergy INTEGER(10),
                     Comments VARCHAR(1024),
